requirements.txt
Flask
openai
stripe
python-dotenv